refine gemm tuning scripts
